+++ src/calendar-input/calendar-input.jsx
@@ -111,25 +111,56 @@ class CalendarInput extends React.Component {
         className: Type.string,
         /** Тема компонента */
         theme: Type.oneOf(['alfa-on-color', 'alfa-on-white']),
-        /** Обработчик установки фокуса на компонент */
+        /**
+         * Обработчик установки фокуса на компонент
+         * @param {React.FocusEvent} event
+         */
         onFocus: Type.func,
-        /** Обработчик снятия фокуса с компонента */
+        /**
+         * Обработчик снятия фокуса с компонента
+         * @param {React.FocusEvent} event
+         */
         onBlur: Type.func,
-        /** Обработчик установки фокуса на поле ввода */
+        /**
+         * Обработчик установки фокуса на поле ввода
+         * @param {React.FocusEvent} event
+         */
         onInputFocus: Type.func,
-        /** Обработчик снятия фокуса с поля ввода */
+        /**
+         * Обработчик снятия фокуса с поля ввода
+         * @param {React.FocusEvent} event
+         */
         onInputBlur: Type.func,
-        /** Обработчик ввода даты в текстовом поле */
+        /**
+         * Обработчик ввода даты в текстовом поле
+         * @param {string} value
+         */
         onInputChange: Type.func,
-        /** Обработчик выбора даты в календаре */
+        /**
+         * Обработчик выбора даты в календаре
+         * @param {string} formattedValue
+         */
         onCalendarChange: Type.func,
-        /** Обрабочик изменения даты в календаре */
+        /**
+         * Обрабочик изменения даты в календаре
+         * @param {string} formattedValue
+         * @param {number} value
+         */
         onChange: Type.func,
-        /** Обработчик события нажатия на клавишу в момент, когда фокус находится на компоненте */
+        /**
+         * Обработчик события нажатия на клавишу в момент, когда фокус находится на компоненте
+         * @param {React.KeyboardEvent} event
+         */
         onKeyDown: Type.func,
-        /** Обработчик события нажатия на клавишу клавиатуры в момент, когда фокус находится в календаре */
+        /**
+         * Обработчик события нажатия на клавишу клавиатуры в момент, когда фокус находится в календаре
+         * @param {React.KeyboardEvent} event
+         */
         onCalendarKeyDown: Type.func,
-        /** Обработчик события нажатия на клавишу клавиатуры в момент, когда фокус находится на текстовом поле */
+        /**
+         * Обработчик события нажатия на клавишу клавиатуры в момент, когда фокус находится на текстовом поле
+         * @param {React.KeyboardEvent} event
+         */
         onInputKeyDown: Type.func
     };
 
@@ -238,65 +269,68 @@ class CalendarInput extends React.Component {
         return (
             <span
                 className={ cn({ width: this.props.width }) }
-                { ...wrapperProps }
             >
-                <Mq
-                    query='--small-only'
-                    touch={ true }
-                    onMatchChange={ this.handleMqMatchChange }
+                <span
+                    { ...wrapperProps }
                 >
-                    {
-                        this.canBeNative() &&
-                        <input
-                            ref={ (nativeCalendarTarget) => {
-                                this.nativeCalendarTarget = nativeCalendarTarget;
-                            } }
-                            { ...commonProps }
-                            { ...nativeProps }
-                            className={ cn('native-control') }
-                            type='date'
-                            value={ changeDateFormat(value, CUSTOM_DATE_FORMAT, NATIVE_DATE_FORMAT) }
-                            onBlur={ this.handleNativeInputBlur }
-                            onChange={ this.handleNativeInputChange }
-                            onFocus={ this.handleNativeInputFocus }
-                        />
-                    }
-                </Mq>
-                <Input
-                    ref={ (customCalendarTarget) => {
-                        this.customCalendarTarget = customCalendarTarget;
-                    } }
-                    { ...commonProps }
-                    className={ cn('custom-control') }
-                    disabledAttr={ this.isNativeInput() || this.isMobilePopup() }
-                    focused={ this.state.isInputFocused || this.state.isCalendarFocused }
-                    mask='11.11.1111'
-                    size={ this.props.size }
-                    type='tel'
-                    pattern='[0-9]*'
-                    label={ this.props.label }
-                    placeholder={ this.props.placeholder }
-                    hint={ this.props.hint }
-                    error={ this.props.error }
-                    value={ value }
-                    width={ this.props.width }
-                    id={ this.props.id }
-                    name={ this.props.name }
-                    leftAddons={ this.props.leftAddons }
-                    rightAddons={ this.props.rightAddons }
-                    onBlur={ this.handleCustomInputBlur }
-                    onChange={ this.handleCustomInputChange }
-                    onFocus={ this.handleCustomInputFocus }
-                    onKeyDown={ this.handleInputKeyDown }
-                    icon={
-                        this.props.withIcon &&
-                        <IconButton onClick={ this.handleIconButtonClick }>
-                            <IconCalendar
-                                size={ this.props.size }
+                    <Mq
+                        query='--small-only'
+                        touch={ true }
+                        onMatchChange={ this.handleMqMatchChange }
+                    >
+                        {
+                            this.canBeNative() &&
+                            <input
+                                ref={ (nativeCalendarTarget) => {
+                                    this.nativeCalendarTarget = nativeCalendarTarget;
+                                } }
+                                { ...commonProps }
+                                { ...nativeProps }
+                                className={ cn('native-control') }
+                                type='date'
+                                value={ changeDateFormat(value, CUSTOM_DATE_FORMAT, NATIVE_DATE_FORMAT) }
+                                onBlur={ this.handleNativeInputBlur }
+                                onChange={ this.handleNativeInputChange }
+                                onFocus={ this.handleNativeInputFocus }
                             />
-                        </IconButton>
-                    }
-                />
+                        }
+                    </Mq>
+                    <Input
+                        ref={ (customCalendarTarget) => {
+                            this.customCalendarTarget = customCalendarTarget;
+                        } }
+                        { ...commonProps }
+                        className={ cn('custom-control') }
+                        disabledAttr={ this.isNativeInput() || this.isMobilePopup() }
+                        focused={ this.state.isInputFocused || this.state.isCalendarFocused }
+                        mask='11.11.1111'
+                        size={ this.props.size }
+                        type='tel'
+                        pattern='[0-9]*'
+                        label={ this.props.label }
+                        placeholder={ this.props.placeholder }
+                        hint={ this.props.hint }
+                        error={ this.props.error }
+                        value={ value }
+                        width={ this.props.width }
+                        id={ this.props.id }
+                        name={ this.props.name }
+                        leftAddons={ this.props.leftAddons }
+                        rightAddons={ this.props.rightAddons }
+                        onBlur={ this.handleCustomInputBlur }
+                        onChange={ this.handleCustomInputChange }
+                        onFocus={ this.handleCustomInputFocus }
+                        onKeyDown={ this.handleInputKeyDown }
+                        icon={
+                            this.props.withIcon &&
+                            <IconButton onClick={ this.handleIconButtonClick }>
+                                <IconCalendar
+                                    size={ this.props.size }
+                                />
+                            </IconButton>
+                        }
+                    />
+                </span>
                 { this.renderPopup(cn, value, Popup) }
             </span>
         );